refactor(service): move linger guards to dedicated Linux-only module
Splits the linger / encrypted-home machinery out of service.ts and
platform.ts into setup/linger.ts. setupSystemd dynamic-imports it so its
body (and the findmnt / fscrypt / marker-dir detection) never evaluates
on macOS or on the WSL nohup fallback path. Pulls setupLinux and
setupSystemd to async so the dynamic import awaits cleanly.

Also folds in two follow-up fixes from PR #2681 review:

- Recovery command in the encrypted-home warning was hardcoded to
  systemctl --user start nanoclaw, but the real unit name is
  nanoclaw-v2-<slug>. A user who copy-pasted hit
  Unit nanoclaw.service not found. manageUserLinger now takes
  unitName as its first argument and interpolates it into the
  warning body; setupSystemd passes the unitName it already has.
  The regression test asserts the real per-install unit name
  appears in the warning, not the static "nanoclaw" string.

- findmnt and fscrypt probes switched from execSync(string) to
  execFileSync(file, args). JSON.stringify does not escape $ or
  backticks, so a hostile homeDir was a (low-likelihood) shell
  injection surface. execFileSync runs without a shell.

Comments added: fscrypt token-match note pinning the verified CLI
version (v0.3.4); ~/.Private false-positive note for post-migration
hosts.

No user-visible behaviour change vs the previous commit aside from
the corrected recovery command. typecheck + tests pass (336/336,
including the four manageUserLinger regression cases).

Refs #2680, #2681.
